// packages/avifil32/types/Avifil32.ts
import type { Pointer } from 'bun:ffi';

export type { BOOL, DWORD, HANDLE, HRESULT, HWND, INT, INT_PTR, LONG, LPCSTR, LPCWSTR, LPSTR, LPVOID, LPWSTR, NULL, UINT, ULONG } from '@bun-win32/core';

export const AVISTREAMREAD_CONVENIENT = -1;

export enum AviCompressFlags {
  AVICOMPRESSF_DATARATE = 0x0000_0002,
  AVICOMPRESSF_INTERLEAVE = 0x0000_0001,
  AVICOMPRESSF_KEYFRAMES = 0x0000_0004,
  AVICOMPRESSF_VALID = 0x0000_0008,
}

export enum AviFileCaps {
  AVIFILECAPS_ALLKEYFRAMES = 0x0000_0010,
  AVIFILECAPS_CANREAD = 0x0000_0001,
  AVIFILECAPS_CANWRITE = 0x0000_0002,
  AVIFILECAPS_NOCOMPRESSION = 0x0000_0020,
}

export enum AviFileInfoFlags {
  AVIFILEINFO_COPYRIGHTED = 0x0002_0000,
  AVIFILEINFO_HASINDEX = 0x0000_0010,
  AVIFILEINFO_ISINTERLEAVED = 0x0000_0100,
  AVIFILEINFO_MUSTUSEINDEX = 0x0000_0020,
  AVIFILEINFO_WASCAPTUREFILE = 0x0001_0000,
}

export enum AviGetFrameFlags {
  AVIGETFRAMEF_BESTDISPLAYFMT = 0x0000_0001,
}

export enum AviIndexFlags {
  AVIIF_COMPUSE = 0x0fff_0000,
  AVIIF_FIRSTPART = 0x0000_0020,
  AVIIF_KEYFRAME = 0x0000_0010,
  AVIIF_LASTPART = 0x0000_0040,
  AVIIF_LIST = 0x0000_0001,
  AVIIF_MIDPART = 0x0000_0060,
  AVIIF_NOTIME = 0x0000_0100,
}

export enum AviStreamInfoFlags {
  AVISTREAMINFO_DISABLED = 0x0000_0001,
  AVISTREAMINFO_FORMATCHANGES = 0x0001_0000,
}

export enum FindSampleFlags {
  FIND_ANY = 0x0000_0020,
  FIND_DIR = 0x0000_000f,
  FIND_FORMAT = 0x0000_0040,
  FIND_FROM_START = 0x0000_0008,
  FIND_INDEX = 0x0000_4000,
  FIND_KEY = 0x0000_0010,
  FIND_LENGTH = 0x0000_1000,
  FIND_NEXT = 0x0000_0001,
  FIND_OFFSET = 0x0000_2000,
  FIND_POS = 0x0000_0000,
  FIND_PREV = 0x0000_0004,
  FIND_RET = 0x0000_f000,
  FIND_SIZE = 0x0000_3000,
  FIND_TYPE = 0x0000_00f0,
}

export enum OpenFileFlags {
  OF_CREATE = 0x0000_1000,
  OF_READ = 0x0000_0000,
  OF_READWRITE = 0x0000_0002,
  OF_SHARE_DENY_NONE = 0x0000_0040,
  OF_SHARE_EXCLUSIVE = 0x0000_0010,
  OF_WRITE = 0x0000_0001,
}

export enum StreamType {
  streamtypeAUDIO = 0x7364_7561,
  streamtypeMIDI = 0x7364_696d,
  streamtypeTEXT = 0x7374_7874,
  streamtypeVIDEO = 0x7364_6976,
}

export type AVISAVECALLBACK = Pointer;
export type LPAVICOMPRESSOPTIONS = Pointer;
export type LPAVIFILEINFO = Pointer;
export type LPAVIFILEINFOA = Pointer;
export type LPAVIFILEINFOW = Pointer;
export type LPAVISTREAMINFO = Pointer;
export type LPAVISTREAMINFOA = Pointer;
export type LPAVISTREAMINFOW = Pointer;
export type LPBITMAPINFOHEADER = Pointer;
export type LPCLSID = Pointer;
export type LPCTSTR = Pointer;
export type LPLONG = Pointer;
export type LPTSTR = Pointer;
export type PAVIFILE = bigint;
export type PAVISTREAM = bigint;
export type PGETFRAME = bigint;